support `startDebugging`
